+++ examples/svelte/.prettierrc.js
@@ -2,4 +2,6 @@ module.exports = {
   arrowParens: 'avoid',
   singleQuote: true,
   trailingComma: 'all',
+  plugins: ['prettier-plugin-svelte'],
+  overrides: [{ files: '*.svelte', options: { parser: 'svelte' } }],
 };