+++ examples/svelte/screens/TabsDemoScreen.svelte
@@ -18,7 +18,11 @@
   ><TabScreen
     name="Home"
     component={TabHomeScreen}
-    options={{ tabBarLabel: 'Home', tabBarIcon: '🏠', tabBarActiveTintColor: LINE_COLOR.structure }}
+    options={{
+      tabBarLabel: 'Home',
+      tabBarIcon: '🏠',
+      tabBarActiveTintColor: LINE_COLOR.structure,
+    }}
   /><TabScreen
     name="Search"
     component={TabSearchScreen}